Create getCurrentDate function
