+++ src/components/Menu/CurrentDate/CurrentDate.tsx
@@ -1,15 +1,10 @@
-import { Container } from './CurrentDate.styled';
+import { CurrentDateContainer } from './CurrentDate.styled';
+import getCurrentDate from './helpers/getCurrentDate';
 
 const CurrentDate = () => {
-  const newDate = new Date();
-  const dateDay = newDate.getDate();
-  const dateMonth = newDate.getMonth() + 1;
-  const dateYear = newDate.getFullYear();
+  const { dateTime, currentDate } = getCurrentDate();
 
-  const dateTime = `${dateYear}-${dateMonth}-${dateDay}`;
-  const currentDate = `${dateDay}.${dateMonth}.${dateYear}`;
-
-  return <Container dateTime={dateTime}>{currentDate}</Container>;
+  return <CurrentDateContainer dateTime={dateTime}>{currentDate}</CurrentDateContainer>;
 };
 
 export default CurrentDate;
